moved some scripts because sr_visualization couldn't find them after installation
